+++ Source/C++/Core/Ap4Descriptor.cpp
@@ -49,8 +49,8 @@ AP4_Descriptor::Inspect(AP4_AtomInspector& inspector)
     AP4_FormatString(name, sizeof(name), "[Descriptor:%02x]", m_ClassId);
     char info[64];
     AP4_FormatString(info, sizeof(info), "size=%d+%d",
-                     GetHeaderSize(),
-                     m_PayloadSize);
+                     (int)GetHeaderSize(),
+                     (int)m_PayloadSize);
     inspector.StartElement(name, info);
     inspector.EndElement();
